+++ packages/gatsby-plugin-sharp/src/trace-svg.js
@@ -146,7 +146,7 @@ const createMemoizedFunctions = () => {
   )
 }
 
-// This is very hacky, but memozied function are pretty tricky to spy on
+// This is very hacky, but memoized function are pretty tricky to spy on
 // in tests ;(
 createMemoizedFunctions()
 exports.createMemoizedFunctions = () => {