+++ packages/gatsby/src/schema/schema.js
@@ -6,6 +6,8 @@ const {
   defaultFieldResolver,
   assertValidName,
   parse,
+  GraphQLNonNull,
+  GraphQLList,
 } = require(`graphql`)
 const {
   ObjectTypeComposer,
@@ -21,7 +23,10 @@ const report = require(`gatsby-cli/lib/reporter`)
 const { addNodeInterfaceFields } = require(`./types/node-interface`)
 const { addInferredType, addInferredTypes } = require(`./infer`)
 const { findOne, findManyPaginated } = require(`./resolvers`)
-const { processFieldExtensions } = require(`./extensions`)
+const {
+  processFieldExtensions,
+  internalExtensionNames,
+} = require(`./extensions`)
 const { getPagination } = require(`./types/pagination`)
 const { getSortInput } = require(`./types/sort`)
 const { getFilterInput } = require(`./types/filter`)
@@ -31,8 +36,9 @@ const buildSchema = async ({
   schemaComposer,
   nodeStore,
   types,
-  thirdPartySchemas,
   typeMapping,
+  fieldExtensions,
+  thirdPartySchemas,
   typeConflictReporter,
   parentSpan,
 }) => {
@@ -40,8 +46,9 @@ const buildSchema = async ({
     schemaComposer,
     nodeStore,
     types,
-    thirdPartySchemas,
     typeMapping,
+    fieldExtensions,
+    thirdPartySchemas,
     typeConflictReporter,
     parentSpan,
   })
@@ -55,6 +62,7 @@ const rebuildSchemaWithSitePage = async ({
   schemaComposer,
   nodeStore,
   typeMapping,
+  fieldExtensions,
   typeConflictReporter,
   parentSpan,
 }) => {
@@ -69,6 +77,7 @@ const rebuildSchemaWithSitePage = async ({
   await processTypeComposer({
     schemaComposer,
     typeComposer,
+    fieldExtensions,
     nodeStore,
     parentSpan,
   })
@@ -85,6 +94,7 @@ const updateSchemaComposer = async ({
   nodeStore,
   types,
   typeMapping,
+  fieldExtensions,
   thirdPartySchemas,
   typeConflictReporter,
   parentSpan,
@@ -107,6 +117,7 @@ const updateSchemaComposer = async ({
       processTypeComposer({
         schemaComposer,
         typeComposer,
+        fieldExtensions,
         nodeStore,
         parentSpan,
       })
@@ -119,11 +130,17 @@ const updateSchemaComposer = async ({
 const processTypeComposer = async ({
   schemaComposer,
   typeComposer,
+  fieldExtensions,
   nodeStore,
   parentSpan,
 }) => {
   if (typeComposer instanceof ObjectTypeComposer) {
-    await processFieldExtensions({ schemaComposer, typeComposer, parentSpan })
+    await processFieldExtensions({
+      schemaComposer,
+      typeComposer,
+      fieldExtensions,
+      parentSpan,
+    })
     if (typeComposer.hasInterface(`Node`)) {
       await addNodeInterfaceFields({ schemaComposer, typeComposer, parentSpan })
       await addResolvers({ schemaComposer, typeComposer, parentSpan })
@@ -153,6 +170,7 @@ const addTypes = ({ schemaComposer, types, parentSpan }) => {
         })
       } catch (error) {
         reportParsingError(error)
+        return
       }
       parsedTypes.forEach(type => {
         processAddedType({
@@ -177,6 +195,7 @@ const addTypes = ({ schemaComposer, types, parentSpan }) => {
         if (schemaComposer.has(typeName)) {
           const typeComposer = schemaComposer.get(typeName)
           mergeTypes({
+            schemaComposer,
             typeComposer,
             type,
             plugin,
@@ -200,6 +219,7 @@ const addTypes = ({ schemaComposer, types, parentSpan }) => {
       if (schemaComposer.has(typeName)) {
         const typeComposer = schemaComposer.get(typeName)
         mergeTypes({
+          schemaComposer,
           typeComposer,
           type: typeOrTypeDef,
           plugin,
@@ -220,6 +240,7 @@ const addTypes = ({ schemaComposer, types, parentSpan }) => {
 }
 
 const mergeTypes = ({
+  schemaComposer,
   typeComposer,
   type,
   plugin,
@@ -237,7 +258,7 @@ const mergeTypes = ({
     if (isNamedTypeComposer(type)) {
       typeComposer.extendExtensions(type.getExtensions())
     }
-    addExtensions({ typeComposer, plugin, createdFrom })
+    addExtensions({ schemaComposer, typeComposer, plugin, createdFrom })
     return true
   } else {
     report.warn(
@@ -268,12 +289,17 @@ const processAddedType = ({
   }
   schemaComposer.addSchemaMustHaveType(typeComposer)
 
-  addExtensions({ typeComposer, plugin, createdFrom })
+  addExtensions({ schemaComposer, typeComposer, plugin, createdFrom })
 
   return typeComposer
 }
 
-const addExtensions = ({ typeComposer, plugin, createdFrom }) => {
+const addExtensions = ({
+  schemaComposer,
+  typeComposer,
+  plugin,
+  createdFrom,
+}) => {
   typeComposer.setExtension(`createdFrom`, createdFrom)
   typeComposer.setExtension(`plugin`, plugin ? plugin.name : null)
 
@@ -314,6 +340,57 @@ const addExtensions = ({ typeComposer, plugin, createdFrom }) => {
           typeComposer.setFieldExtension(fieldName, name, args)
         })
       }
+
+      // Validate field extension args. `graphql-compose` already checks the
+      // type of directive args in `parseDirectives`, but we want to check
+      // extensions provided with type builders as well. Also, we warn if an
+      // extension option was provided which does not exist in the field
+      // extension definition.
+      const fieldExtensions = typeComposer.getFieldExtensions(fieldName)
+      const typeName = typeComposer.getTypeName()
+      Object.keys(fieldExtensions)
+        .filter(name => !internalExtensionNames.includes(name))
+        .forEach(name => {
+          const args = fieldExtensions[name]
+          try {
+            const definition = schemaComposer.getDirective(name)
+
+            // Handle `defaultValue` when not provided as directive
+            definition.args.forEach(({ name, defaultValue }) => {
+              if (args[name] === undefined && defaultValue !== undefined) {
+                args[name] = defaultValue
+              }
+            })
+
+            Object.keys(args).forEach(arg => {
+              const argumentDef = definition.args.find(
+                ({ name }) => name === arg
+              )
+              if (!argumentDef) {
+                report.error(
+                  `Field extension \`${name}\` on \`${typeName}.${fieldName}\` ` +
+                    `has invalid argument \`${arg}\`.`
+                )
+                return
+              }
+              const value = args[arg]
+              try {
+                validate(argumentDef.type, value)
+              } catch (error) {
+                report.error(
+                  `Field extension \`${name}\` on \`${typeName}.${fieldName}\` ` +
+                    `has argument \`${arg}\` with invalid value "${value}". ` +
+                    error.message
+                )
+              }
+            })
+          } catch (error) {
+            report.error(
+              `Field extension \`${name}\` on \`${typeName}.${fieldName}\` ` +
+                `is not available.`
+            )
+          }
+        })
     })
   }
 
@@ -418,7 +495,7 @@ const addSetFieldsOnGraphQLNodeTypeFields = ({
             nodes: nodeStore.getNodesByType(typeName),
           },
           traceId: `initial-setFieldsOnGraphQLNodeType`,
-          parentSpan: parentSpan,
+          parentSpan,
         })
         if (result) {
           // NOTE: `setFieldsOnGraphQLNodeType` only allows setting
@@ -547,7 +624,7 @@ const addCustomResolveFunctions = async ({ schemaComposer, parentSpan }) => {
     schema: intermediateSchema,
     createResolvers,
     traceId: `initial-createResolvers`,
-    parentSpan: parentSpan,
+    parentSpan,
   })
 }
 
@@ -695,6 +772,7 @@ const parseTypes = ({
 
       // Merge the parsed type with the original
       mergeTypes({
+        schemaComposer,
         typeComposer,
         type: parsedType,
         plugin,
@@ -728,7 +806,6 @@ const reportParsingError = error => {
   const { message, source, locations } = error
 
   if (source && locations && locations.length) {
-    const report = require(`gatsby-cli/lib/reporter`)
     const { codeFrameColumns } = require(`@babel/code-frame`)
 
     const frame = codeFrameColumns(
@@ -761,3 +838,19 @@ const isNamedTypeComposer = type =>
   type instanceof EnumTypeComposer ||
   type instanceof InterfaceTypeComposer ||
   type instanceof UnionTypeComposer
+
+const validate = (type, value) => {
+  if (type instanceof GraphQLNonNull) {
+    if (value == null) {
+      throw new Error(`Expected non-null field value.`)
+    }
+    return validate(type.ofType, value)
+  } else if (type instanceof GraphQLList) {
+    if (!Array.isArray(value)) {
+      throw new Error(`Expected array field value.`)
+    }
+    return value.map(v => validate(type.ofType, v))
+  } else {
+    return type.parseValue(value)
+  }
+}