+++ packages/gatsby-plugin-google-gtag/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.1.0...gatsby-plugin-google-gtag@1.1.1) (2019-07-02)
+
+**Note:** Version bump only for package gatsby-plugin-google-gtag
+
 # [1.1.0](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-plugin-google-gtag/compare/gatsby-plugin-google-gtag@1.0.17...gatsby-plugin-google-gtag@1.1.0) (2019-06-20)
 
 **Note:** Version bump only for package gatsby-plugin-google-gtag